requirements.txt
pandas>=1.3.0
numpy>=1.21.0
yfinance>=0.1.70
ta-lib>=0.4.24
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.28.0
openpyxl>=3.0.0
streamlit>=1.22.0
plotly>=5.13.0